requirements.txt
streamlit==1.32.2
langchain==0.3.23
langchain-core==0.3.51
langchain-groq==0.3.2
groq==0.22.0
langchain-openai==0.3.12
langgraph==0.3.26
chromadb==0.4.24
python-dotenv==1.0.1
pydantic==2.11.3
faiss-cpu==1.10.0
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.28.1
langchain-community==0.3.21